record_pets_project/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-api-key==2.3.0
python-dotenv==1.0.0
Pillow==9.0.0
drf-spectacular==0.27.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
drf-extra-fields==3.7.0